pyproject.toml
[tool.poetry]
name = "python-word"
version = "0.1.0"
description = ""
authors = ["Daisuke Takahashi <stumble.on.the.stair@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-docx = "^0.8.10"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
